src/main/webapp/app/entities/marca/update/marca-update.component.spec.ts
jest.mock('@angular/router');

import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { of, Subject } from 'rxjs';

import { MarcaService } from '../service/marca.service';
import { IMarca, Marca } from '../marca.model';

import { MarcaUpdateComponent } from './marca-update.component';

describe('Component Tests', () => {
  describe('Marca Management Update Component', () => {
    let comp: MarcaUpdateComponent;
    let fixture: ComponentFixture<MarcaUpdateComponent>;
    let activatedRoute: ActivatedRoute;
    let marcaService: MarcaService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [MarcaUpdateComponent],
        providers: [FormBuilder, ActivatedRoute],
      })
        .overrideTemplate(MarcaUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(MarcaUpdateComponent);
      activatedRoute = TestBed.inject(ActivatedRoute);
      marcaService = TestBed.inject(MarcaService);

      comp = fixture.componentInstance;
    });

    describe('ngOnInit', () => {
      it('Should update editForm', () => {
        const marca: IMarca = { id: 456 };

        activatedRoute.data = of({ marca });
        comp.ngOnInit();

        expect(comp.editForm.value).toEqual(expect.objectContaining(marca));
      });
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Marca>>();
        const marca = { id: 123 };
        jest.spyOn(marcaService, 'update').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ marca });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.next(new HttpResponse({ body: marca }));
        saveSubject.complete();

        // THEN
        expect(comp.previousState).toHaveBeenCalled();
        expect(marcaService.update).toHaveBeenCalledWith(marca);
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Marca>>();
        const marca = new Marca();
        jest.spyOn(marcaService, 'create').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ marca });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.next(new HttpResponse({ body: marca }));
        saveSubject.complete();

        // THEN
        expect(marcaService.create).toHaveBeenCalledWith(marca);
        expect(comp.isSaving).toEqual(false);
        expect(comp.previousState).toHaveBeenCalled();
      });

      it('Should set isSaving to false on error', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Marca>>();
        const marca = { id: 123 };
        jest.spyOn(marcaService, 'update').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ marca });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.error('This is an error!');

        // THEN
        expect(marcaService.update).toHaveBeenCalledWith(marca);
        expect(comp.isSaving).toEqual(false);
        expect(comp.previousState).not.toHaveBeenCalled();
      });
    });
  });
});
